Add: `getCanUserSendMetatransactions` generator helper
